4032 - useCanSeeUserActivities move to store/user
