+++ src/client/pages/CountryHome/FraHome/hooks/useSections.tsx
@@ -5,8 +5,7 @@ import { AssessmentNames } from 'meta/assessment'
 import { SectionNames } from 'meta/routes'
 
 import { useAssessment, useCycle } from 'client/store/assessment'
-import { useUser } from 'client/store/user'
-import { useCanSeeUserActivities } from 'client/hooks/useCanSeeUserActivities'
+import { useCanSeeUserActivities, useUser } from 'client/store/user'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
 import Collaborators from 'client/pages/CountryHome/Collaborators'
 import RecentActivity from 'client/pages/CountryHome/FraHome/RecentActivity'
