+++ src/client/pages/CountryHome/PanEuropeanHome/PanEuropeanHome.tsx
@@ -9,8 +9,7 @@ import { Areas } from 'meta/area'
 import { MessageTopicType, Topics } from 'meta/messageCenter'
 import { SectionNames } from 'meta/routes'
 
-import { useUser } from 'client/store/user'
-import { useCanSeeUserActivities } from 'client/hooks/useCanSeeUserActivities'
+import { useCanSeeUserActivities, useUser } from 'client/store/user'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
 import MessageButton from 'client/components/MessageButton'
 import ButtonDownloadDashboard from 'client/pages/CountryHome/FraHome/ButtonDownloadDashboard'
