+++ src/client/pages/CountryHome/PanEuropeanHome/hooks/useSections.tsx
@@ -4,9 +4,8 @@ import { SectionNames } from 'meta/routes'
 import { Users } from 'meta/user'
 
 import { useCycle } from 'client/store/assessment'
-import { useUser } from 'client/store/user'
+import { useCanSeeUserActivities, useUser } from 'client/store/user'
 import { useCountryIso } from 'client/hooks'
-import { useCanSeeUserActivities } from 'client/hooks/useCanSeeUserActivities'
 import Collaborators from 'client/pages/CountryHome/Collaborators'
 import RecentActivity from 'client/pages/CountryHome/FraHome/RecentActivity'
 import Repository from 'client/pages/CountryHome/Repository'
