+++ src/client/store/user/hooks/index.ts
@@ -130,3 +130,26 @@ export const useCanViewReview = (sectionName: string) => {
   const canView = Authorizer.canViewReview({ country, cycle, section, user })
   return !print && !isDataLocked && canView
 }
+
+/**
+ * React hook to determine whether given user has access to view user activities (eg. Messaging, Recent activity, etc.)
+ *
+ * @param user - The user
+ * @returns boolean indicating whether the user can view user activities
+ *
+ * @example
+ * const user = useUser();
+ * const canSeeActivities = useCanSeeUserActivities(user);
+ *
+ * if (!canSeeActivities) {
+ *   // Hide activities UI
+ * }
+ */
+export const useCanSeeUserActivities = (user: User) => {
+  const { countryIso } = useCountryRouteParams()
+  const cycle = useCycle()
+
+  const rolesAllowedToView = Users.getRolesAllowedToView({ user, countryIso, cycle })
+
+  return rolesAllowedToView.length > 0
+}
