+++ src/client/store/user/index.ts
@@ -1,6 +1,7 @@
 export {
   useCanEdit,
   useCanEditCycleData,
+  useCanSeeUserActivities,
   useIsCountryRepositoryEditable,
   useIsEditTableDataEnabled,
   useIsGlobalRepositoryEditable,
